chore(models): Cleans up prints from GPT-2 Flex and moves ProjectedAdaptiveLogSoftmax to model_utils.
